requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0

# Database
motor>=3.3.1
pymongo>=4.5.0
redis>=5.0.1
rq>=1.15.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
itsdangerous>=2.1.2
authlib>=1.2.1

# AI and LLM
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-anthropic>=0.1.1
langchain-google-genai>=0.0.5
openai>=1.3.0
google-generativeai>=0.3.1
anthropic>=0.8.1

# Web crawling
crawl4ai>=0.1.0

# Media processing
moviepy>=1.0.3
opencv-python>=4.8.1.78
ffmpeg-python>=0.2.0
boto3>=1.28.64

# Monitoring and observability
sentry-sdk>=1.32.0
prometheus-client>=0.17.1
prometheus-fastapi-instrumentator>=6.1.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.0

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0
tenacity>=8.2.3
aiohttp>=3.8.6 